Created multitrajresult module
