+++ doc/apidoc/classes.rst
@@ -231,7 +231,7 @@ Solver Options and Results
     :inherited-members:
     :exclude-members: add_processor, add
 
-.. autoclass:: qutip.solver.result.MultiTrajResult
+.. autoclass:: qutip.solver.multitrajresult.MultiTrajResult
     :members:
     :inherited-members:
     :exclude-members: add_processor, add, add_end_condition
@@ -240,11 +240,11 @@ Solver Options and Results
     :show-inheritance:
     :members:
 
-.. autoclass:: qutip.solver.result.McResult
+.. autoclass:: qutip.solver.multitrajresult.McResult
     :show-inheritance:
     :members:
 
-.. autoclass:: qutip.solver.result.NmmcResult
+.. autoclass:: qutip.solver.multitrajresult.NmmcResult
     :show-inheritance:
     :members:
